Remove sveltekit
