+++ frontend/package.json
@@ -7,9 +7,8 @@
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"prepare": "svelte-kit sync || echo ''",
-		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
 		"format": "prettier --write .",
 		"lint": "prettier --check . && eslint .",
 		"test:unit": "vitest",
@@ -22,8 +21,6 @@
 		"@internationalized/date": "^3.12.0",
 		"@lucide/svelte": "^0.561.0",
 		"@playwright/test": "^1.58.2",
-		"@sveltejs/adapter-static": "^3.0.10",
-		"@sveltejs/kit": "^2.53.4",
 		"@sveltejs/vite-plugin-svelte": "^6.2.4",
 		"@tailwindcss/vite": "^4.2.1",
 		"@types/node": "^25.3.5",
@@ -50,6 +47,7 @@
 		"typescript": "^5.9.3",
 		"typescript-eslint": "^8.56.1",
 		"vite": "^7.3.1",
+		"vite-plugin-singlefile": "^2.1.0",
 		"vitest": "^4.0.18"
 	},
 	"engines": {
@@ -64,10 +62,6 @@
 		"onlyBuiltDependencies": [
 			"@tailwindcss/oxide",
 			"esbuild"
-		],
-		"overrides": {
-			"cookie@<0.7.0": ">=0.7.0",
-			"js-yaml@<4.1.1": ">=4.1.1"
-		}
+		]
 	}
 }
