+++ frontend/src/app.d.ts
@@ -1,13 +1,4 @@
-// See https://svelte.dev/docs/kit/types#app.d.ts
-// for information about these interfaces
 declare global {
-  namespace App {
-    // interface Error {}
-    // interface Locals {}
-    // interface PageData {}
-    // interface PageState {}
-    // interface Platform {}
-  }
   interface Navigator {
     userAgentData?: {
       mobile?: boolean;
