+++ frontend/src/lib/components/Layout/Header.svelte
@@ -7,12 +7,7 @@
 >
   <div class="flex flex-1 flex-row items-center justify-between space-x-2 px-4 py-2">
     <div class="flex items-center space-x-4">
-      <a
-        href="/"
-        class="overflow-hidden select-none lg:ml-0! lg:w-auto"
-        data-sveltekit-preload-data="hover"
-        aria-label="OpenShock"
-      >
+      <a href="/" class="overflow-hidden select-none lg:ml-0! lg:w-auto" aria-label="OpenShock">
         <img
           class="pointer-events-none inline-block h-6 sm:h-10"
           src="/logo.svg"
