--- frontend/src/routes/+layout.ts
@@ -1,1 +1,0 @@
-export const prerender = true;
