+++ frontend/svelte.config.js
@@ -1,20 +1,12 @@
-import adapter from '@sveltejs/adapter-static';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
-/** @type {import('@sveltejs/kit').Config} */
+/** @type {import('@sveltejs/vite-plugin-svelte').SvelteConfig} */
 const config = {
   preprocess: vitePreprocess(),
 
   compilerOptions: {
     runes: true,
     modernAst: true,
-  },
-
-  kit: {
-    adapter: adapter(),
-    output: {
-      bundleStrategy: 'inline',
-    },
   },
 };
